Refactor repository structure with database locking
