+++ stratz_scraper/__init__.py
@@ -1,0 +1,3 @@
+from .web.app import create_app
+
+__all__ = ["create_app"]
