+++ stratz_scraper/web/static/css/styles.css
