+++ stratz_scraper/web/static/js/app.js
