+++ stratz_scraper/web/templates/index.html
